+++ packages/core-cairo/CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## Unreleased
 
+- Add ERC721Enumerable. ([#391](https://github.com/OpenZeppelin/contracts-wizard/pull/391))
+
 - **Breaking changes**:
   - Use OpenZeppelin Contracts for Cairo v0.17.0. ([#396](https://github.com/OpenZeppelin/contracts-wizard/pull/396))
-
 
 ## 0.15.0 (2024-09-19)